+++ src/collection/map.js
@@ -1,9 +1,10 @@
-define(['../lang/isObject', '../object/values', '../array/map'], function (isObject, values, arrMap) {
+define(['../lang/isObject', '../object/values', '../array/map', '../function/makeIterator_'], function (isObject, values, arrMap, makeIterator) {
 
     /**
      * Map collection values, returns Array.
      */
     function map(list, callback, thisObj) {
+        callback = makeIterator(callback);
         // list.length to check array-like object, if not array-like
         // we simply map all the object values
         if( isObject(list) && list.length == null ){