+++ src/object/map.js
@@ -1,10 +1,11 @@
-define(['./forOwn'], function(forOwn) {
+define(['./forOwn', '../function/makeIterator_'], function(forOwn, makeIterator) {
 
     /**
      * Creates a new object where all the values are the result of calling
      * `callback`.
      */
     function mapValues(obj, callback, thisObj) {
+        callback = makeIterator(callback);
         var output = {};
         forOwn(obj, function(val, key, obj) {
             output[key] = callback.call(thisObj, val, key, obj);