+++ src/object/some.js
@@ -1,9 +1,10 @@
-define(['./forOwn'], function(forOwn) {
+define(['./forOwn', '../function/makeIterator_'], function(forOwn, makeIterator) {
 
     /**
      * Object some
      */
     function some(obj, callback, thisObj) {
+        callback = makeIterator(callback);
         var result = false;
         forOwn(obj, function(val, key) {
             if (callback.call(thisObj, val, key, obj)) {